requirements.txt
google-genai>=1.0.0
requests>=2.31.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
tweepy>=4.14.0
playwright>=1.40.0
